move common type structs to test_common.go
